Added a potential method for partitioning test data from training data
This method uses a histogram to split the data into bins.  A user-defined proportion of each bin is sampled.  Then the data is split back into testing and training data.

Notably, the distribution of mass in the dataset is concentrated in the first few bins.  As such, I split the data into 3 different subsets based on two arbitrary borders.  This allows a better representation of data in various bins.